code_cifar10/requirements.txt
absl-py>=0.5.0
numpy>=1.13.3
tensor2tensor>=1.9.0
# tensorflow or tensorflow-gpu
